incr decr tests
